[qtcontacts-sqlite] Allow suppression of sync contacts change signal
This commit ensures that the syncContactsChanged signal is not emitted
for the sync target of the sync adapter which causes the change.
That signal exists to tell the sync adapter that it may need to
perform a sync due to a change on the local device; if the change was
caused by that sync adapter, then we don't need to tell it about the
change, obviously.

This helps prevent sync triggering cycles.